Change dt from float to double

+++ triangleCollision/pba/base/Solver.C
@@ -5,7 +5,7 @@
 using namespace std;
 using namespace pba;
 
-void SolverBase::updatePos(float dt)
+void SolverBase::updatePos(double dt)
 {
     for (size_t i = 0; i < DS->nb(); ++i)
     {
@@ -14,7 +14,7 @@ void SolverBase::updatePos(float dt)
     }
 }
 
-void SolverBase::updateVel(float dt)
+void SolverBase::updateVel(double dt)
 {
     for (size_t i = 0; i < DS->nb(); ++i)
     {
@@ -25,7 +25,7 @@ void SolverBase::updateVel(float dt)
 }
 
 
-void LeapFrogSolver::_updateDS(float dt)
+void LeapFrogSolver::_updateDS(double dt)
 {
     // S_lf(dt) = S_x(dt/2)S_v(dt)S_x(dt/2)
     LeapFrogSolver::updatePos(dt / 2);
@@ -34,12 +34,12 @@ void LeapFrogSolver::_updateDS(float dt)
 }
 
 
-void SixOrderSolver::_updateDS(float dt)
+void SixOrderSolver::_updateDS(double dt)
 {
-    float a = float(1.0 / (4.0 - std::pow(4, 1.0/3.0)));
-    float b = 1 - 4 * a;
-    float dt_a = a * dt;
-    float dt_b = b * dt;
+    double a = 1.0 / (4.0 - std::pow(4, 1.0/3.0));
+    double b = 1 - 4 * a;
+    double dt_a = a * dt;
+    double dt_b = b * dt;
 
     LeapFrogSolver::updateDS(dt_a);
     LeapFrogSolver::updateDS(dt_a);
